[algs] Add a float overload of hb_clamp_to
Call sites feeding roundf/floorf/ceilf results into hb_clamp_to
promoted float to double implicitly, tripping -Wdouble-promotion under
clang. The overload widens explicitly — float cannot represent all
integer bounds exactly, so the clamp itself stays in double — and the
two glyf extents assignments that promote outside hb_clamp_to get
explicit casts. No behavior change.

+++ src/OT/glyf/glyf.hh
@@ -357,8 +357,8 @@ struct glyf_accelerator_t
 	  return;
 	}
 	{
-	  double x_bearing = roundf (min_x);
-	  double y_bearing = roundf (max_y);
+	  double x_bearing = (double) roundf (min_x);
+	  double y_bearing = (double) roundf (max_y);
 	  extents->x_bearing = hb_clamp_to<hb_position_t> (x_bearing);
 	  extents->width = hb_clamp_to<hb_position_t> ((double) roundf (max_x) - x_bearing);
 	  extents->y_bearing = hb_clamp_to<hb_position_t> (y_bearing);
